requirements.txt
python-binance>=1.0.0
datetime
matplotlib>=3.5.0
numpy>=1.20.0
pandas>=1.3.0
seaborn>=0.11.0
streamlit>=1.10.0
scipy>=1.7.0
